uml: fuse: Skip test since it needs qcow2.

+++ fuse/test-fuse-umount-race.sh
@@ -33,6 +33,11 @@ if [ ! -w /dev/fuse ]; then
     exit 77
 fi
 
+if [ "$(../fish/guestfish get-backend)" = "uml" ]; then
+    echo "$0: skipping test because uml backend does not support qcow2"
+    exit 77
+fi
+
 rm -f test.qcow2 test-copy.qcow2 test.pid
 rm -rf mp
